requirements.txt
groq>=0.4.1,<1.0.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.0,<3.0